requirements.txt
# Telegram bots
python-telegram-bot==22.*
telethon>=1.36.0,<2.0

# HTTP client utilities
httpx>=0.27.0

# Environment configuration
python-dotenv>=1.0.1

# Settings / validation helpers
pydantic>=2.8

